4.21 (4.12 puuttuu)

+++ controllers/blogs.js
@@ -64,12 +64,31 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
     try {
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+        if (!request.token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        const blog = await Blog.findById(request.params.id)
+
+        if (blog.user.toString() !== decodedToken.id.toString()) {
+            return response.status(401).json({ 
+                error: 'You can only remove blogs added by you'
+             })
+        }
+
+
         await Blog.findByIdAndRemove(request.params.id)
 
         response.status(204).end()
     } catch (exception) {
-        console.log(exception)
-        response.status(400).send({ error: 'malformatted id' })
+        if (exception.name === 'JsonWebTokenError') {
+            response.status(401).json({ error: exception.message })
+        } else {
+            console.log(exception)
+            response.status(400).send({ error: 'malformatted id' })
+        }
     }
 })
